очередной коммит

+++ task/task10/task1001/Solution.java
@@ -11,7 +11,7 @@ public class Solution {
         int b = (byte) a + 46;
         byte c = (byte) (a * b);
         double f = (char) 1234.15;
-        long d = (short) (a + f / c + b);
+        long d = (int) (a + f / c + b);
         System.out.println(d);
     }
 }
